电子章

+++ src/main/java/com/smart/bracelet/model/po/assets/Manner1.java
@@ -30,7 +30,7 @@ public class Manner1 {
      *支付现金
      */
     @NotNull(message = "支付现金不能为空")
-    private Long cash;
+    private Float cash;
 
     /**
      * 甲方法人代表
